+++ src/main/java/com/ssafy/soljigi/base/error/AppException.java
@@ -1,4 +1,4 @@
-package com.ssafy.soljigi.user.error;
+package com.ssafy.soljigi.base.error;
 
 import lombok.Getter;
