+++ src/main/java/com/ssafy/soljigi/base/error/exception/InvalidTokenException.java
@@ -1,0 +1,10 @@
+package com.ssafy.soljigi.base.error.exception;
+
+import com.ssafy.soljigi.base.error.AppException;
+import com.ssafy.soljigi.base.error.ErrorCode;
+
+public class InvalidTokenException extends AppException {
+	public InvalidTokenException() {
+		super(ErrorCode.INVALID_TOKEN);
+	}
+}
