--- src/main/java/com/ssafy/soljigi/user/error/exception/InvalidTokenException.java
@@ -1,10 +1,0 @@
-package com.ssafy.soljigi.user.error.exception;
-
-import com.ssafy.soljigi.user.error.AppException;
-import com.ssafy.soljigi.user.error.ErrorCode;
-
-public class InvalidTokenException extends AppException {
-	public InvalidTokenException() {
-		super(ErrorCode.INVALID_TOKEN);
-	}
-}
